Move the model classes
