+++ src/app/canlines/canlines.component.ts
@@ -1,7 +1,8 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { plainToClass } from 'class-transformer';
 import { CanSerialService } from '../can-serial-service.service';
-import { CanData } from '../CanLinesModel';
+import { CanData } from './shared/candata.model';
+import { CanLine } from './shared/canline.model';
 
 @Component({
   selector: 'app-canlines',
@@ -13,10 +14,15 @@ export class CanlinesComponent implements OnInit {
   constructor(private canSerialService : CanSerialService) { }
 
   CanData : CanData = new CanData();
+  CurrentKeyToMonitor : string = "";
 
   ngOnInit(): void {
     this.canSerialService.OnNewModel.subscribe((value) => {
       this.CanData = plainToClass(CanData, value);
     });
+  }
+
+  getDataToMonitor () : CanLine[] {
+    return this.CanData.getCanLinesForKey(this.CurrentKeyToMonitor);
   }
 }
