+++ src/app/canlines/shared/dictionary.model.ts
@@ -1,0 +1,3 @@
+export interface Dictionary<T> {
+  [Key: string]: T;
+}
